🔐 Add Order policy and refactor Controllers

+++ app/Http/Controllers/CartController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Services\CartService;
+use App\Models\Cart;
 use App\Models\CartItem;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -18,34 +19,58 @@ class CartController extends Controller
         $this->cartService = $cartService;
     }
 
-    public function index()
+    public function index(Cart $cart)
     {
+        $cart = $this->cartService->getOrCreateUserCart();
+
+        $this->authorize('index', $cart);
+
         $items = $this->cartService->getCartItems();
+        
         return response()->json($items);
     }
 
     public function store(AddCartItemRequest $request)
     {
-        $item = $this->cartService->addItem($request->validated());
+        $cart = $this->cartService->getOrCreateUserCart();
+        
+        $this->authorize('addItem', $cart);
+
+        $validated = $request->validated();
+
+        $item = $this->cartService->addItem($validated);
+
         return response()->json($item, 201);
     }
 
     public function update(UpdateCartRequest $request, CartItem $cartItem)
     {
+        $cart = $this->cartService->getOrCreateUserCart();  
+
+        $this->authorize('update', $cart);
+
         $validated = $request->validated();
+
         $item = $this->cartService->updateItem($cartItem, $validated['quantity']);
+        
         return response()->json($item);
     }
 
     public function destroy(CartItem $cartItem)
     {
+        $this->authorize('delete', $cartItem);
+
         $this->cartService->removeItem($cartItem);
+        
         return response()->json(null, 204);
     }
 
     public function clear()
     {
+        $this->authorize('clear', CartItem::class);
+
         $this->cartService->clearCart();
+        
         return response()->json(null, 204);
     }
 }
